feature-fin-1279

+++ model/src/main/java/ir/demisco/cfs/model/dto/response/FinancialAccountTurnOverRecordsResponse.java
@@ -6,11 +6,11 @@ public class FinancialAccountTurnOverRecordsResponse {
     private Date documentDate;
     private String documentNumber;
     private String description;
-    private Double debitAmount;
-    private Double creditAmount;
-    private Double remainDebit;
-    private Double remainCredit;
-    private Double remainAmount;
+    private Long debitAmount;
+    private Long creditAmount;
+    private Long remainDebit;
+    private Long remainCredit;
+    private Long remainAmount;
     private Long recordType;
 
     private FinancialAccountTurnOverSummarizeResponse financialAccountTurnOverSummarizeModel;
@@ -40,43 +40,43 @@ public class FinancialAccountTurnOverRecordsResponse {
         this.description = description;
     }
 
-    public Double getDebitAmount() {
+    public Long getDebitAmount() {
         return debitAmount;
     }
 
-    public void setDebitAmount(Double debitAmount) {
+    public void setDebitAmount(Long debitAmount) {
         this.debitAmount = debitAmount;
     }
 
-    public Double getCreditAmount() {
+    public Long getCreditAmount() {
         return creditAmount;
     }
 
-    public void setCreditAmount(Double creditAmount) {
+    public void setCreditAmount(Long creditAmount) {
         this.creditAmount = creditAmount;
     }
 
-    public Double getRemainDebit() {
+    public Long getRemainDebit() {
         return remainDebit;
     }
 
-    public void setRemainDebit(Double remainDebit) {
+    public void setRemainDebit(Long remainDebit) {
         this.remainDebit = remainDebit;
     }
 
-    public Double getRemainCredit() {
+    public Long getRemainCredit() {
         return remainCredit;
     }
 
-    public void setRemainCredit(Double remainCredit) {
+    public void setRemainCredit(Long remainCredit) {
         this.remainCredit = remainCredit;
     }
 
-    public Double getRemainAmount() {
+    public Long getRemainAmount() {
         return remainAmount;
     }
 
-    public void setRemainAmount(Double remainAmount) {
+    public void setRemainAmount(Long remainAmount) {
         this.remainAmount = remainAmount;
     }
 
@@ -125,27 +125,27 @@ public class FinancialAccountTurnOverRecordsResponse {
             return this;
         }
 
-        public Builder debitAmount(Double debitAmount) {
+        public Builder debitAmount(Long debitAmount) {
             financialAccountTurnOverRecordsResponse.setDebitAmount(debitAmount);
             return this;
         }
 
-        public Builder creditAmount(Double creditAmount) {
+        public Builder creditAmount(Long creditAmount) {
             financialAccountTurnOverRecordsResponse.setCreditAmount(creditAmount);
             return this;
         }
 
-        public Builder remainDebit(Double remainDebit) {
+        public Builder remainDebit(Long remainDebit) {
             financialAccountTurnOverRecordsResponse.setRemainDebit(remainDebit);
             return this;
         }
 
-        public Builder remainCredit(Double remainCredit) {
+        public Builder remainCredit(Long remainCredit) {
             financialAccountTurnOverRecordsResponse.setRemainCredit(remainCredit);
             return this;
         }
 
-        public Builder remainAmount(Double remainAmount) {
+        public Builder remainAmount(Long remainAmount) {
             financialAccountTurnOverRecordsResponse.setRemainAmount(remainAmount);
             return this;
         }
